Tidy up integrations, use contracts
