+++ app/Integrations/GitHub/Releases.php
@@ -9,12 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace CachetHQ\Cachet\Integrations;
+namespace CachetHQ\Cachet\Integrations\GitHub;
 
+use CachetHQ\Cachet\Integrations\Contracts\Releases as ReleasesContract;
 use GuzzleHttp\Client;
 use Illuminate\Contracts\Cache\Repository;
 
-class Releases
+class Releases implements ReleasesContract
 {
     /**
      * The default url.
